backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
httpx
pydantic-settings
PyJWT[crypto]
pytest
pytest-asyncio
pytest-mock
openai>=1.0.0